Added a keyword to trigger auto-pull

+++ auto_update.sh
@@ -2,8 +2,9 @@
 
 while true
 do
-	nc -l 4200
-	cd ~/stalker_sama && git pull
-	./stop.sh
-	./run.sh | tee stalker.log
+	if nc -l 4200 | grep -iq "\[prod ready\]"; then
+		cd ~/stalker_sama && git pull
+		./stop.sh
+		./run.sh | tee stalker.log
+	fi
 done
